third_party/blink/web_tests/accessibility/menu-list-open.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<select id="menulist">
  <option id="item0">Alicia</option>
  <option id="item1">Peter</option>
  <option id="item2">Kalinda</option>
</select>
<script>
async_test(function(t)
{
    var menulist = document.getElementById("menulist");
    menulist.selectedIndex = 0;
    menulist.focus();

    var eventCount = 0;
    accessibilityController.addNotificationListener(function listener(element, notification) {
        if (notification == "ActiveDescendantChanged") {
            assert_equals(element.role, "AXRole: AXMenuListPopup");
            ++eventCount;
        }
        if (notification == "MenuListItemSelected") {
            assert_equals(element.role, "AXRole: AXMenuListOption");
            assert_equals(element.name, "Alicia");
            ++eventCount;
        }

        if (eventCount == 2)
            t.done();
    });

    var axMenuList = accessibilityController.accessibleElementById("menulist");
    eventSender.mouseMoveTo(axMenuList.x + 10, axMenuList.y + 10);
    eventSender.mouseDown();
    eventSender.mouseUp();
}, "menu list fires correct events on open");
</script>
